feat(chat): reuse key
