+++ chat/src/logic.rs
@@ -1,5 +1,6 @@
 mod auto_update;
 mod gossip;
+mod key;
 mod message;
 
 use auto_update::AutoUpdater;
@@ -9,7 +10,7 @@ use iroh_gossip::{
     net::{Event, GossipEvent, GossipSender},
     proto::TopicId,
 };
-use iroh_net::key::{PublicKey, SecretKey};
+use iroh_net::key::PublicKey;
 use key::get_or_create_secret_key;
 use message::Message;
 use std::{
@@ -28,39 +29,6 @@ pub struct Logic {
 
 const TOPIC_ID: &str = "63fi4am3m2uu47ylikwbnkac4nyqiookbnlmcxkosqqugpg2ayja";
 const PEER_IDS: [&str; 1] = ["4h6tmz5id4yh4f6jwpdi5s6a42z4tf2ulmzcqx2o337572cbutvq"];
-
-mod key {
-    use iroh_net::key::SecretKey;
-    use std::{fs, path::PathBuf};
-    use tracing::info;
-
-    fn key_file_path() -> PathBuf {
-        std::env::current_exe()
-            .expect("failed to get current executable path")
-            .parent()
-            .expect("failed to get parent directory")
-            .join("user_key.bin")
-    }
-
-    pub fn get_or_create_secret_key() -> SecretKey {
-        let path = key_file_path();
-        match fs::read(&path) {
-            Ok(bytes) if bytes.len() == 32 => {
-                let bytes: [u8; 32] = bytes.try_into().unwrap();
-                info!("loaded existing user key");
-                SecretKey::from_bytes(&bytes)
-            }
-            _ => {
-                info!("generating new user key");
-                let key = SecretKey::generate();
-                if let Err(e) = fs::write(&path, key.to_bytes()) {
-                    info!("failed to save user key: {}", e);
-                }
-                key
-            }
-        }
-    }
-}
 
 impl Logic {
     pub fn new() -> Self {
